requirements.txt
flask
flask-cors
psycopg2-binary
python-aliexpress-api
python-dotenv
apscheduler
gunicorn

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
